sink(ticdc): remove kafka producer
