+++ cdc/sinkv2/eventsink/mq/kafka_dml_sink.go
@@ -20,7 +20,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/cdc/sink/mq/dispatcher"
-	"github.com/pingcap/tiflow/cdc/sink/mq/producer/kafka"
 	"github.com/pingcap/tiflow/cdc/sinkv2/eventsink/mq/dmlproducer"
 	"github.com/pingcap/tiflow/cdc/sinkv2/util"
 	"github.com/pingcap/tiflow/pkg/config"
@@ -65,7 +64,7 @@ func NewKafkaDMLSink(
 	}()
 
 	// adjust the option configuration before creating the kafka client
-	if err = kafka.AdjustOptions(ctx, adminClient, options, topic); err != nil {
+	if err = pkafka.AdjustOptions(ctx, adminClient, options, topic); err != nil {
 		return nil, cerror.WrapError(cerror.ErrKafkaNewProducer, err)
 	}
